Fixed bug for algorithm for finding free time.
Algorithm was not returning overlapping friend events past a certain point even though the events were within the time range specified by the user.
This was because there was no start and end timestamp that was passed into the function running the algorithm, and the behavior of the algorithm was undefined.
As well, in the algorithm, the last event was excluded by accident due to misunderstanding the index vs the number of events.

Updated Firestore rules to only read and write if user is authenticated. Next step is to specify rules for reading and writing to specific collections.